% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_params_c2t.R
\name{c2t_module}
\alias{c2t_module}
\alias{c2t_ui}
\alias{c2t_server}
\title{Conditions to Treatments Module}
\usage{
c2t_ui(id)

c2t_server(id, params, redraw_c2t, counter, popn_subgroup, conditions)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params, redraw_c2t, popn_subgroup, conditions}{reactive objects passed in from the params page}

\item{counter}{the counter object from the params page}
}
\description{
A shiny module that renders all of the content for the c2t section in the params page
}
