% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{combined_plot}
\alias{combined_plot}
\title{Combined Plot}
\usage{
combined_plot(model_output, treatment, params)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{treatment}{a name of a treatment to filter by}

\item{params}{the current `params` object used to model the data}
}
\value{
a plotly chart
}
\description{
Generates a plot that shows the actual demand data, along with the suppressed activty figures, and the model output.
}
