% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{combined_plot_data}
\alias{combined_plot_data}
\title{Combined Plot Data}
\usage{
combined_plot_data(model_output, treatment, params)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{treatment}{a name of a treatment to filter by}

\item{params}{the current `params` object used to model the data}
}
\value{
a summarised version of \code{model_output}
}
\description{
Prepares the data for \code{combined_plot()}
}
