% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{comma}
\alias{comma}
\title{Comma Format}
\usage{
comma(x)
}
\arguments{
\item{x}{a number to convert to comma format}
}
\value{
a string representation of the input in comma format
}
\description{
A simple version of scales::comma
}
