% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counter.R
\docType{class}
\name{Counter-class}
\alias{Counter-class}
\title{Counter Class}
\description{
Counter Class
}
\section{Fields}{

\describe{
\item{\code{value}}{the current value of the counter, should not be accessed directly}
}}

\section{Methods}{

\describe{
\item{\code{get()}}{increments the counter and return's the new value}
}}

