% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_create_graph.R
\name{create_graph}
\alias{create_graph}
\title{Create Graph}
\usage{
create_graph(
  model_output,
  groups = unique(model_output$group),
  conditions = unique(model_output$condition),
  treatments = unique(model_output$treatment)
)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{groups}{a vector of population groups to include in the graph, defaults to all groups}

\item{conditions}{a vector of conditions to include in the graph, defaults to all conditions}

\item{treatments}{a vector of treatments to include in the graph, defaults to all treatments}
}
\value{
a plotly chart
}
\description{
Generates a network graph of the flows from population groups, to conditions, to treatments, as generated by the
model.
}
