% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_demand.R
\name{demand_module}
\alias{demand_module}
\alias{demand_ui}
\title{Demand Module}
\usage{
demand_ui(id)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params}{reactive object passed in from the main server}
}
\description{
A shiny module that renders all of the content for the demand page.
}
