% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demand.R
\name{demand_plot}
\alias{demand_plot}
\alias{demand_plot_data}
\title{Demand Plot}
\usage{
demand_plot(model_output, appointments, treatment)

demand_plot_data(model_output, appointments, treatment)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{appointments}{output from \code{get_appointments()}}

\item{treatment}{a name of a treatment to filter by}
}
\value{
\code{demand_plot()}: a plotly chart

\code{demand_plot_data()}: a summarised version of \code{model_output}
}
\description{
Generates a plot that shows the demand generated by the model.
}
