% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_output.R
\name{download_output}
\alias{download_output}
\title{Download Model Output}
\usage{
download_output(model_output, params)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{params}{the current `params` object used to model the data}
}
\value{
a function that accepts a file name to save the results to
}
\description{
Download the model output as a csv in order to be used with other applications
}
