% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_report.R
\name{download_report}
\alias{download_report}
\title{Download Report}
\usage{
download_report(model_output, params, services)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{params}{the current `params` object used to model the data}

\item{services}{the services to generate this report for - ignored if "All" is selected for \code{download_choice}}
}
\value{
a function that accepts a file name to save the report to
}
\description{
Generates a pdf report to download for the current services
}
