% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extract_params_from_excel.R
\name{extract_params_from_excel}
\alias{extract_params_from_excel}
\title{Extract Params From Excel File}
\usage{
extract_params_from_excel(raw_data_path)
}
\arguments{
\item{raw_data_path}{the filepath where the Excel file is located}
}
\value{
`param` object to use in the model
}
\description{
Extracts the params object to use in the model from an Excel file
}
