% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fct_reorder.R
\name{fct_reorder}
\alias{fct_reorder}
\title{Reorder factor levels by sorting along another variable}
\usage{
fct_reorder(.f, .x, .fun = median, ..., .desc = FALSE)
}
\arguments{
\item{.f}{A factor (or character vector).}

\item{.x}{The levels of f are reordered so that the values of .fun(.x) are in ascending order.}

\item{.fun}{summary function. It should take one vector}

\item{...}{Other arguments passed on to .fun. A common argument is na.rm = TRUE.}

\item{.desc}{Order in descending order?}
}
\description{
fct_reorder() is useful for 1d displays where the factor is mapped to position; fct_reorder2() for 2d displays where
the factor is mapped to a non-position aesthetic. last2() and first2() are helpers for fct_reorder2(); last2() finds
the last value of y when sorted by x; first2() finds the first value.
}
\details{
Reimplements \code{\link[forcats:fct_reorder]{forcats::fct_reorder()}}
}
