% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_g2c.R
\name{g2c_module}
\alias{g2c_module}
\alias{g2c_ui}
\alias{g2c_server}
\title{Group to Conditions Module}
\usage{
g2c_ui(id)

g2c_server(id, params, redraw_g2c, redraw_c2t, counter, popn_subgroup)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params, redraw_g2c, redraw_c2t, popn_subgroup}{reactive objects passed in from the params page}

\item{counter}{the counter object from the params page}
}
\description{
A shiny module that renders all of the content for the g2c section in the params page
}
