% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_all_conditions.R
\name{get_all_conditions}
\alias{get_all_conditions}
\title{Get All Conditions}
\usage{
get_all_conditions(params)
}
\arguments{
\item{params}{the current params used in the model}
}
\value{
a character vector of the condition names
}
\description{
Extracts the names of the conditions from the current params
}
