% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_output_helpers.R
\name{get_appointments}
\alias{get_appointments}
\title{Get Appointments}
\usage{
get_appointments(params)
}
\arguments{
\item{params}{the current `params` object used to model the data}
}
\value{
a tibble of treatments and average monthly appointments
}
\description{
Extracts the average month appointments each treatment generates per month from the params
}
