% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_output_helpers.R
\name{get_model_output}
\alias{get_model_output}
\title{Get model output}
\usage{
get_model_output(models, start_month)
}
\arguments{
\item{models}{the model output list}

\item{start_month}{what is month 0? should be a date object that is the first of the month}
}
\value{
a tibble
}
\description{
Takes models list and returns a dataframe of the model results
}
