% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_helpers.R
\name{get_model_params}
\alias{get_model_params}
\title{Get Model Params}
\usage{
get_model_params(params)
}
\arguments{
\item{params}{the current `params` object used to model the data}
}
\value{
a matrix of the parameters for the Systems Dynamics model
}
\description{
Converts the params object into a form that is usable by \code{run_model()}
}
