% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_helpers.R
\name{get_model_potential_functions}
\alias{get_model_potential_functions}
\title{Get Model Potential Functions}
\usage{
get_model_potential_functions(params, months)
}
\arguments{
\item{params}{the current `params` object used to model the data}

\item{months}{the number of months to run the model for}
}
\value{
a list of functions for each of the population groups
}
\description{
Takes the current params and generates the functions that simulate when the populations enter the model
}
