% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{half_life_factor}
\alias{half_life_factor}
\title{Half Life Factor}
\usage{
half_life_factor(t, p = 0.5)
}
\arguments{
\item{t}{A number, the time it takes for an initial value to decay to `p`\% of the original value}

\item{p}{A number, the percentage to reduce to (defaults to 50\%)}
}
\value{
A value in \[0, 1\] that is the percentage that would remain after 1 time period elapses
}
\description{
Calculates from a time parameter and a percentage a value for the amount of a value that remains after
1 period of time
}
