% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_help_popups.R
\name{help_popups}
\alias{help_popups}
\title{Help Popups}
\usage{
help_popups(page)
}
\arguments{
\item{page}{the page to load help file for, stored in inst/app/data/\{\{page\}\}_help.json}
}
\description{
Adds help files to a page
}
