% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_home.R
\name{home_module}
\alias{home_module}
\alias{home_ui}
\title{Home Module}
\usage{
home_ui(id)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}
}
\description{
A shiny module that renders all of the content for the home page.
}
