% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_md_to_tags.R
\name{md_to_tags}
\alias{md_to_tags}
\title{Markdown to Tags}
\usage{
md_to_tags(file)
}
\arguments{
\item{file}{the file to read and convert to html tags}
}
\value{
a shiny.tag.list object
}
\description{
Takes a markdown file and converts it to a HTML tags object that can be used in a Shiny UI
}
