% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_output_helpers.R
\name{model_totals}
\alias{model_totals}
\title{Model Totals}
\usage{
model_totals(model_output, type, treatment)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{type}{a character vector of the "type" to total}

\item{treatment}{a name of a treatment to filter by}
}
\value{
a comma-formatted number for the total for that type/treatment
}
\description{
Get the total value generated by the model for a given treatment. Can be used to calculate for referrals, treatments
etc.
}
