% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_params.R
\name{params_module}
\alias{params_module}
\alias{params_ui}
\alias{params_server}
\title{Params Module}
\usage{
params_ui(id)

params_server(id, params, model_output, upload_event, params_file_path)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params, model_output}{reactive objects passed in from the main server}

\item{upload_event}{a reactiveVal that is updated when a file is uploaded}

\item{params_file_path}{a reactiveVal that contains the path to the current params file}
}
\description{
A shiny module that renders all of the content for the params page.
}
