% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactive_changes.R
\name{reactive_changes}
\alias{reactive_changes}
\title{Reactive Changes}
\usage{
reactive_changes(expr)
}
\arguments{
\item{expr}{an expression that gets the values to observe whether they are changing}
}
\value{
a reactiveVal
}
\description{
Create a reactive that is only updated when the expression returns a different vector to what is currently stored
}
