% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reduce_condition_pcnts.R
\name{reduce_condition_pcnts}
\alias{reduce_condition_pcnts}
\title{Reduce Condition Percents}
\usage{
reduce_condition_pcnts(conditions, current_conditions)
}
\arguments{
\item{conditions}{the conditions (from params$groups[[.x]]$conditions)}

\item{current_conditions}{the names of conditions to reduce by, initially all except the condition that has been
changed}
}
\value{
the altered conditions
}
\description{
For each group, the sum of the "pcnt" value for each condition must not exceed 100%. This function recursively
iterates reducing the percents equally of the other conditions.
}
\details{
If at any point the amount that we need to reduce (equally) by exceeds the smallest value, then we reduce all by the
smallest group, then remove the smallest group and iterate.
}
