% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{referrals_plot}
\alias{referrals_plot}
\title{Referrals Plot}
\usage{
referrals_plot(model_output, treatment)
}
\arguments{
\item{treatment}{a name of a treatment to filter by}

\item{model_ouput}{output from \code{run_model()} and \code{get_model_output()}}
}
\value{
a plotly chart
}
\description{
Generates a plot that shows the referrals generated by the model.
}
