% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_referrals_plot.R
\name{referrals_plot_data}
\alias{referrals_plot_data}
\title{Referrals Plot Data}
\usage{
referrals_plot_data(model_output, treatment)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{treatment}{a name of a treatment to filter by}
}
\value{
a summarised version of \code{model_output}
}
\description{
Prepares the data for \code{referrals_plot()}
}
