% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{replace_bootstrap_cols}
\alias{replace_bootstrap_cols}
\title{Replace bootstrap grid columns}
\usage{
replace_bootstrap_cols(
  x,
  from = c(".", "xs", "sm", "md", "lg", "xl"),
  to = c("xs", "sm", "md", "lg", "xl")
)
}
\arguments{
\item{x}{an object we want to iterate through, initialy should be a shiny.tag.list}

\item{from}{the column type we want to replace, defaults to "." (all)}

\item{to}{the column type we want to replace}
}
\description{
By default shiny uses col-sm-* classes. We want to be able to replace these with a different column type, e.g.
col-xl-*. This function recursively iterates through shiny UI elements and replaces the classes of shiny.tag objects.
}
