% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_results.R
\name{results_module}
\alias{results_module}
\alias{results_ui}
\alias{results_server}
\title{Results Module}
\usage{
results_ui(id)

results_server(id, model_output, params)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{model_output, params}{reactive objects passed in from the main server}
}
\description{
A shiny module that renders all of the content for the results page.
}
