% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_helpers.R
\name{run_all_models}
\alias{run_all_models}
\title{Run All Models}
\usage{
run_all_models(params, months, sim_time)
}
\arguments{
\item{params}{the current `params` object used to model the data}

\item{months}{an integer of the number of months to run the simulation for}

\item{sim_time}{a numeric for the time interval to run the model at, e.g. every 1/5th of a month}
}
\value{
the output of \code{run_model()}
}
\description{
Run's the model for all of the population group
}
