% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{run_model}
\alias{run_model}
\title{Run Model}
\usage{
run_model(params, new_potential, simtime = seq(0, 18, by = 1/30))
}
\arguments{
\item{params}{the value returned from \code{get_model_params()} from the current `params`}

\item{new_potential}{the value returned from \code{get_model_potential_functions()} from the current `params`}

\item{simtime}{the points at which we should try to solve the function at. Defaults to 0 to 18 (months), solving at
intervals of 1 / 30 (e.g. each day in the month)}
}
\value{
a tibble of the results of running the model
}
\description{
Performs the Systems Dynamics modelling on the given parameter values.
}
