% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_helpers.R
\name{run_single_model}
\alias{run_single_model}
\title{Run Single Model}
\usage{
run_single_model(params, groups, months, sim_time)
}
\arguments{
\item{params}{the current `params` object used to model the data}

\item{groups}{a character vector of the population group to model}

\item{months}{an integer of the number of months to run the simulation for}

\item{sim_time}{a numeric for the time interval to run the model at, e.g. every 1/5th of a month}
}
\value{
the output of \code{run_model()} for the selected population group
}
\description{
Run's the model for a single population group
}
