% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_setup.R
\name{setup_module}
\alias{setup_module}
\alias{setup_ui}
\alias{setup_server}
\title{Setup Module}
\usage{
setup_ui(id)

setup_server(id, params)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params}{reactive object passed in from the main server}
}
\description{
A shiny module that renders all of the content for the setup page.
}
