% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subpopulation_curve_plot.R
\name{subpopulation_curve_plot}
\alias{subpopulation_curve_plot}
\title{Subpopulation curve plot}
\usage{
subpopulation_curve_plot(curve, size, pcnt)
}
\arguments{
\item{curve}{the curve vector to display}

\item{size}{the base population size}

\item{pcnt}{the percentage of the base population size we are using (0 - 100)}
}
\value{
a plotly chart
}
\description{
displays a quick plot of what the current parameters selected for the subpopulation group will generate
}
