% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_helpers.R
\name{summarise_model_output}
\alias{summarise_model_output}
\title{Referrals Plot}
\usage{
summarise_model_output(model_output, type, treatment)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{treatment}{a name of a treatment to filter by}
}
\value{
a plotly chart
}
\description{
Generates a plot that shows the referrals generated by the model.
}
