% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surge.R
\name{surge_plot_data}
\alias{surge_plot_data}
\title{Surge Plot Data}
\usage{
surge_plot_data(model_output, column)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{column}{the column in `model_output` to calculate the summary for, either "group", "condition" or "treatment"}
}
\value{
a summarised version of \code{model_output}
}
\description{
Generates a plot that shows the amount of people who were referred and treated, and referred but not treated, as
generated by the model.
}
