% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_output_helpers.R
\name{surge_summary}
\alias{surge_summary}
\title{Surge Summary}
\usage{
surge_summary(model_output, column)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{column}{the column in `model_output` to calculate the summary for, either "group", "condition" or "treatment"}
}
\value{
a tibble containing a row per group/condition/treatment and the total value generated by the model
}
\description{
Calculates the total surge generated by the model for the provided column.
}
