% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_surgetab.R
\name{surgetab_module}
\alias{surgetab_module}
\alias{surgetab_ui}
\alias{surgetab_server}
\title{Surge Tab Module}
\usage{
surgetab_ui(id)

surgetab_server(id, model_output, column, title)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{model_output}{reactive object passed in from the main server}

\item{column}{a NSE column name to use to display}

\item{title}{the name to use to describe the column used}
}
\description{
A shiny module that renders all of the content of the surge tabs
}
