% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{treatment_split_plot}
\alias{treatment_split_plot}
\title{Treatment split plot}
\usage{
treatment_split_plot(treatments)
}
\arguments{
\item{treatments}{a named vector of the splits for the treatments}
}
\value{
a plotly chart
}
\description{
generates a bar chart to show the percentage splits for the current treatments
}
