% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_treemap.R
\name{treemap_module}
\alias{treemap_module}
\alias{treemap_ui}
\alias{treemap_server}
\title{Treemap Module}
\usage{
treemap_ui(id)

treemap_server(id, params)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params}{reactive object passed in from the main server}
}
\description{
A shiny module that renders all of the content for the treemap page.
}
