% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemap.R
\name{treemap_plot}
\alias{treemap_plot}
\title{Treemap plot}
\usage{
treemap_plot(params)
}
\arguments{
\item{params}{the current `params` object used to model the data}
}
\value{
a plotly chart
}
\description{
Creates a treemap of the population group sizes at baseline.
}
