% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{where}
\alias{where}
\title{Where predicate}
\usage{
where(fn)
}
\arguments{
\item{fn}{a predicate function that should accept a value and return a logical}
}
\value{
a function that applies \code{fn()} with an input \code{(x)} and optional additional arguments
}
\description{
reimplements tidyselect::where. This is far simpler but will suffice for our needs
}
