% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{demand_plot}
\alias{demand_plot}
\title{Demand Plot}
\usage{
demand_plot(model_output, appointments, treatment)
}
\arguments{
\item{appointments}{output from \code{get_appointments()}}

\item{treatment}{a name of a treatment to filter by}

\item{model_ouput}{output from \code{run_model()} and \code{get_model_output()}}
}
\value{
a plotly chart
}
\description{
Generates a plot that shows the demand generated by the model.
}
