% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_report.R
\name{download_report}
\alias{download_report}
\title{Download Report}
\usage{
download_report(services, model_output, appointments, params)
}
\arguments{
\item{services}{the services to generate this report for}

\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{appointments}{output from \code{get_appointments()}}

\item{params}{the current `params` object used to model the data}
}
\value{
a function that accepts a file name to save the report to
}
\description{
Generates a pdf report to download for the current services
}
