% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_model_output}
\alias{get_model_output}
\title{Get model output}
\usage{
get_model_output(models)
}
\arguments{
\item{models}{the model output list}
}
\value{
a tibble
}
\description{
Takes models list and returns a dataframe of the model results
}
