% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_graph.R
\name{graph_module}
\alias{graph_module}
\alias{graph_ui}
\alias{graph_server}
\title{Graph Module}
\usage{
graph_ui(id)

graph_server(id, params, model_output)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params, model_output}{reactive objects passed in from the main server}
}
\description{
A shiny module that renders all of the content for the graph page.
}
