% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_params.R
\name{params_module}
\alias{params_module}
\alias{params_ui}
\alias{params_server}
\title{Params Module}
\usage{
params_ui(id)

params_server(id, params, model_output)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{params, model_output, }{reactive objects passed in from the main server}
}
\description{
A shiny module that renders all of the content for the params page.
}
