% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{popgroups_plot}
\alias{popgroups_plot}
\title{Population Groups Plot}
\usage{
popgroups_plot(model_output, treatment)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{treatment}{a name of a treatment to filter by}
}
\value{
a plotly chart
}
\description{
Generates a plot that shows the size of the population groups for a given treatment.
}
