% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_demand.R, R/module_results.R
\name{demand_server}
\alias{demand_server}
\alias{results_module}
\alias{results_ui}
\alias{results_server}
\title{Results Module}
\usage{
demand_server(id, params)

results_ui(id)

results_server(id, params, model_output)
}
\arguments{
\item{id}{An ID string that uniquely identifies an instance of this module}

\item{model_output, params}{reactive objects passed in from the main server}
}
\description{
A shiny module that renders all of the content for the results page.
}
