% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{run_model}
\alias{run_model}
\title{Run Model}
\usage{
run_model(params, sim_time)
}
\arguments{
\item{params}{the current `params` object used to model the data}

\item{sim_time}{a numeric for the time interval to run the model at, e.g. every 1/5th of a month}
}
\value{
a tibble of the results of running the model
}
\description{
Performs the Systems Dynamics modelling on the given parameter values.
}
