% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_helpers.R
\name{summarise_model_output}
\alias{summarise_model_output}
\title{Summarise model output}
\usage{
summarise_model_output(model_output, type, treatment)
}
\arguments{
\item{model_output}{output from \code{run_model()} and \code{get_model_output()}}

\item{type}{the row "type" to filter by}

\item{treatment}{a name of a treatment to filter by}
}
\value{
a filtered and summarised version of \code{model_output}
}
\description{
Helper function used to filter results of the model output for use in plots and tables
}
