% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{surge_plot}
\alias{surge_plot}
\title{Surge Plot}
\usage{
surge_plot(data)
}
\arguments{
\item{data}{output from \code{surge_summary()}}
}
\value{
a plotly chart
}
\description{
Generates a plot that shows the amount of people who were referred and treated, and referred but not treated, as
generated by the model.
}
