% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_output_helpers.R
\name{surge_table}
\alias{surge_table}
\title{Surge Table}
\usage{
surge_table(surge_data, group_name)
}
\arguments{
\item{surge_data}{output from \code{surge_summary}}

\item{group_name}{the name of the "group" used, e.g. group, condition, treatment}
}
\value{
a tibble with more descriptive column names
}
\description{
Helper function to take output from \code{surge_summary} and produce a table of data for display in the surge tabs.
}
